treblle_rest_api_hackathon/requirements.txt
argcomplete==2.0.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
blinker==1.4
Brlapi==0.8.4
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
chardet==4.0.0
charset-normalizer==2.1.1
chrome-gnome-shell==0.0.0
click==8.1.3
cloudinary==1.31.0
colorama==0.4.5
colorlog==6.7.0
command-not-found==0.3
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.6
distro==1.7.0
distro-info==1.1+ubuntu1
dj-database-url==1.2.0
Django==4.1.3
django-ckeditor==6.5.1
django-configurations==2.4.1
django-cors-headers==3.13.0
django-countries==7.4.2
django-environ==0.9.0
django-extensions==3.2.1
django-filter==22.1
django-js-asset==2.0.0
django-pagedown==2.2.1
django-password-reset==2.0
django-rest-framework==0.1.0
django-rest-knox==4.2.0
django-rest-passwordreset==1.3.0
django-restframework==0.0.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-spectacular==0.25.1
duplicity==0.8.22
fasteners==0.17.3
filelock==3.8.0
flake8==6.0.0
future==0.18.2
gunicorn==20.1.0
gyp==0.1
html5lib==1.1
httplib2==0.20.4
idna==3.4
importlib-metadata==4.6.4
inflection==0.5.1
itypes==1.2.0
jeepney==0.8.0
Jinja2==3.1.2
jsonschema==4.17.3
keyring==23.8.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.22.0
lxml==4.9.1
mailchimp-transactional==1.0.50
Mako==1.1.3
MarkupSafe==2.1.1
mccabe==0.7.0
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.4
netifaces==0.11.0
nox==2022.11.21
numpy==1.21.5
oauthlib==3.2.2
olefile==0.46
packaging==23.0
paramiko==2.10.4
pexpect==4.8.0
Pillow==9.3.0
pipenv==2022.9.24
platformdirs==2.5.2
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pycairo==1.20.1
pycodestyle==2.10.0
pycparser==2.21
pycryptodomex==3.11.0
pycups==2.0.1
pyflakes==3.0.1
Pygments==2.14.0
PyGObject==3.42.2
PyJWT==2.6.0
PyNaCl==1.5.0
pyparsing==3.0.7
pyrsistent==0.19.3
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-debian==0.1.46
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.6
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
redis==4.4.2
render-cli==0.2.1
reportlab==3.6.11
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
scour==0.38.2
SecretStorage==3.3.3
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
soupsieve==2.3.2
sqlparse==0.4.3
systemd-python==235
tweepy @ git+https://github.com/tweepy/tweepy.git@6fde20d61fd21b06408dedc38f542987cf91c1bc
typing_extensions==4.4.0
tzdata==2022.6
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.13
usb-creator==0.3.7
virtualenv==20.16.5
wadllib==1.3.6
webencodings==0.5.1
whitenoise==6.4.0
xdg==5
xkit==0.0.0
youtube-dl==2021.12.17
zipp==1.0.0
